requirements.txt
pymupdf==1.24.10
pandas==2.2.2
transformers==4.53.2
peft==0.16.0
accelerate==1.8.1
bitsandbytes==0.46.1
datasets==4.0.0
trl==0.19.1
sentence-transformers==5.0.0
protobuf==3.20.*
openai
claudette==0.3.1
vllm==0.9.2
ninja
hydra-core
sentencepiece
einops
kagglehub
wandb
matplotlib
seaborn
python-Levenshtein
pandarallel
lxml